+++ autokeras/__init__.py
@@ -1,5 +1,8 @@
 from autokeras.image.image_supervised import ImageClassifier, ImageRegressor, PortableImageSupervised
 from autokeras.text.text_supervised import TextClassifier, TextRegressor
-from autokeras.object_detector import ObjectDetector
 from autokeras.tabular.tabular_supervised import TabularClassifier, TabularRegressor
+
 from autokeras.net_module import CnnModule, MlpModule
+
+from autokeras.pretrained.object_detector import ObjectDetector
+from autokeras.pretrained.face_detector import FaceDetector
