--- autokeras/object_detection/__init__.py
@@ -1,3 +1,0 @@
-# from .data import *
-# from .layers import *
-# from .utils import *
