--- autokeras/object_detection/data/voc.py
@@ -1,26 +1,0 @@
-"""VOC Dataset Classes
-
-Original author: Francisco Massa
-https://github.com/fmassa/vision/blob/voc_dataset/torchvision/datasets/voc.py
-
-Updated by: Ellis Brown, Max deGroot
-"""
-from .config import HOME
-import os
-import os.path as osp
-import sys
-import torch
-import torch.utils.data as data
-import cv2
-import numpy as np
-from sklearn import model_selection
-
-VOC_CLASSES = ( # always index 0
-    'aeroplane', 'bicycle', 'bird', 'boat',
-    'bottle', 'bus', 'car', 'cat', 'chair',
-    'cow', 'diningtable', 'dog', 'horse',
-    'motorbike', 'person', 'pottedplant',
-    'sheep', 'sofa', 'train', 'tvmonitor')
-
-# note: if you used our download scripts, this should be right
-VOC_ROOT = osp.join(HOME, "object_detection/data/VOCdevkit/")
