--- autokeras/object_detection/data/voc0712.py
@@ -1,28 +1,0 @@
-"""VOC Dataset Classes
-
-Original author: Francisco Massa
-https://github.com/fmassa/vision/blob/voc_dataset/torchvision/datasets/voc.py
-
-Updated by: Ellis Brown, Max deGroot
-"""
-from .config import HOME
-import os.path as osp
-import sys
-import torch
-import torch.utils.data as data
-import cv2
-import numpy as np
-if sys.version_info[0] == 2:
-    import xml.etree.cElementTree as ET
-else:
-    import xml.etree.ElementTree as ET
-
-VOC_CLASSES = ( # always index 0
-    'aeroplane', 'bicycle', 'bird', 'boat',
-    'bottle', 'bus', 'car', 'cat', 'chair',
-    'cow', 'diningtable', 'dog', 'horse',
-    'motorbike', 'person', 'pottedplant',
-    'sheep', 'sofa', 'train', 'tvmonitor')
-
-# note: if you used our download scripts, this should be right
-VOC_ROOT = osp.join(HOME, "object_detection/data/VOCdevkit/")
