--- autokeras/object_detection/layers/__init__.py
@@ -1,2 +1,0 @@
-from .functions import *
-from .modules import *
