--- autokeras/object_detection/layers/functions/__init__.py
@@ -1,5 +1,0 @@
-from .detection import Detect
-from .prior_box import PriorBox
-
-
-__all__ = ['Detect', 'PriorBox']
