--- autokeras/object_detection/layers/modules/__init__.py
@@ -1,3 +1,0 @@
-from .l2norm import L2Norm
-
-__all__ = ['L2Norm']
