--- autokeras/object_detection/utils/__init__.py
@@ -1,1 +1,0 @@
-
